Break apart files.
